Remove Site entity #43
