+++ jdi-light-html-tests/src/main/java/io/github/com/pages/HomePage.java
@@ -6,9 +6,14 @@ import com.epam.jdi.light.elements.pageobjects.annotations.simple.UI;
 import com.epam.jdi.light.ui.html.common.Label;
 import com.epam.jdi.light.ui.html.common.Link;
 import com.epam.jdi.light.ui.html.common.Text;
+import static com.epam.jdi.light.settings.WebSettings.DOMAIN;
 
 public class HomePage extends WebPage {
 	@Css("h3[name='main-title']") public static Label mainTitle;
 	@Css(".main-txt") public static Text jdiText;
 	@UI("['JDI Github']") public static Link githubLink;
+	
+	public static void open() {
+		WebPage.openUrl(DOMAIN + "/index.html");
+	}
 }
