+++ jdi-light-html-tests/src/test/resources/test.properties
@@ -17,3 +17,4 @@ page.check.after.open=NONE
 #run.type=local | remote
 #driver.remote.url=http://localhost:4444/wd/hub
 #screenshot.strategy=on fail | on | off
+page.init.path=src/main/java/io/github/com
